backend/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.4
python-dotenv==1.0.0

# Data processing

scikit-learn==1.8.0
pandas==2.3.1
imbalanced-learn==0.11.0

# ML/AI
xgboost==1.7.6
langchain==0.0.267
langgraph==0.0.15
faiss-cpu==1.10.0
sentence-transformers==2.2.2

# AWS
boto3==1.28.0
botocore==1.31.0

# Database
sqlalchemy==2.0.19
redis==4.6.0

# Utilities
celery==5.3.1
pydantic==2.11.7
joblib==1.3.1
werkzeug==2.3.7